More strict error handling by replacement of "error-strings"
