+++ include/apps/_experiments/OswAppWeather.h
@@ -3,6 +3,7 @@
 #include <osw_hal.h>
 #include <vector>
 #include <OswAppV1.h>
+#include "ArduinoJson.h"
 #include "OswAppWeatherIconPrinter.h"
 
 class OswAppWeather : public OswApp {
@@ -20,6 +21,52 @@ class OswAppWeather : public OswApp {
     virtual void stop() override;
     ~OswAppWeather() {};
   private:
+    class WeatherDecoder {
+      public:
+        WeatherDecoder(const String& input_String);
+        bool strIsValid();
+        time_t getTime();
+        std::vector<OswAppWeather::weather_update_t> getUpdates();
+      private:
+        time_t _str2time(const String& t);
+        int _str2temp(const String& temp);
+        int _str2hum(const String& humidity);
+        int _str2pres(const String& pressure);
+        int _str2wthr(const String& weather);
+        bool in_ok = true;
+        int nUpdates = 0;
+        String in_String;
+    };
+
+    class WeatherParser {
+      public:
+        WeatherParser();
+        std::optional<String> encodeWeather(DynamicJsonDocument& doc);
+      private:
+        int _getWCond(int weather_code);
+        int cnt;
+        std::vector<OswAppWeather::weather_update_t> updates;
+        std::vector<int>clearCode{800};//0
+        std::vector<int>cloudsMin{801};//1
+        std::vector<int>cloudsMed{802};//2
+        std::vector<int>cloudsHigh{803, 804};//3
+        std::vector<int>mist{701};//4
+        std::vector<int>fog{741};//5
+        std::vector<int>snowMin{611, 612, 615, 616};//6
+        std::vector<int>snowMed{600, 613, 601, 620};//7
+        std::vector<int>snowHigh{602, 621, 622};//8
+        std::vector<int>rainMin{500, 300, 301, 302, 310, 311, 312, 313, 314, 321};//9
+        std::vector<int>rainMed{502, 501};//10
+        std::vector<int>rainHigh{503, 504, 511, 520, 521, 522, 531};//11
+        std::vector<int>thunderstorm{200, 201, 210, 211, 231, 230};//12
+        std::vector<int>thunderstormHeavy{202, 212, 221, 232};//13
+        std::vector<int>squallTornado{771, 781};//14
+        //15 ->unknown
+        std::vector<std::vector<int>>weather_conditions{clearCode, cloudsMin, cloudsMed, cloudsHigh, mist, fog, snowMin, snowMed,
+                snowHigh, rainMin, rainMed, rainHigh, thunderstorm,
+                thunderstormHeavy, squallTornado };
+    };
+
     void getW();
     void printW();
     void drawLayout();
