+++ src/apps/_experiments/OswAppWeather.cpp
@@ -8,7 +8,6 @@
 #include <OswAppV1.h>
 #include <osw_hal.h>
 #include "fonts/DS_DIGI12pt7b.h"
-#include "ArduinoJson.h"
 #ifndef OSW_EMULATOR
 #include <HTTPClient.h>
 #endif
@@ -25,24 +24,9 @@
 
 
 
-class WeatherDecoder {
-  public:
-    WeatherDecoder(const String& input_String);
-    bool strIsValid();
-    time_t getTime();
-    std::vector<OswAppWeather::weather_update_t> getUpdates();
-  private:
-    time_t _str2time(const String& t);
-    int _str2temp(const String& temp);
-    int _str2hum(const String& humidity);
-    int _str2pres(const String& pressure);
-    int _str2wthr(const String& weather);
-    bool in_ok = true;
-    int nUpdates = 0;
-    String in_String;
-};
 
-WeatherDecoder::WeatherDecoder(const String& input_String) {
+
+OswAppWeather::WeatherDecoder::WeatherDecoder(const String& input_String) {
     if(input_String.length() <  16 || (input_String.length()%8)!= 0 ) {
         this->in_ok = false;
     }
@@ -50,17 +34,17 @@ WeatherDecoder::WeatherDecoder(const String& input_String) {
     this->nUpdates = (this->in_String.length()-8)/8;
 }
 
-bool WeatherDecoder::strIsValid() {
+bool OswAppWeather::WeatherDecoder::strIsValid() {
     return in_ok;
 }
 
-time_t WeatherDecoder::getTime() {
+time_t OswAppWeather::WeatherDecoder::getTime() {
     String time_str = this->in_String.substring(0,8);
     time_t t = this->_str2time(time_str);
     return t;
 }
 
-std::vector<OswAppWeather::weather_update_t> WeatherDecoder::getUpdates() {
+std::vector<OswAppWeather::weather_update_t> OswAppWeather::WeatherDecoder::getUpdates() {
     OswAppWeather::weather_update_t update;
     String update_str;
     std::vector<OswAppWeather::weather_update_t> updates;
@@ -75,77 +59,47 @@ std::vector<OswAppWeather::weather_update_t> WeatherDecoder::getUpdates() {
     return updates;
 }
 
-time_t WeatherDecoder::_str2time(const String& t) {
+time_t OswAppWeather::WeatherDecoder::_str2time(const String& t) {
     int time = t.toInt();
     time = time * 8;
     time = 2147483647 - time;
     return time;
 }
 
-int WeatherDecoder::_str2temp(const String& temp) {
+int OswAppWeather::WeatherDecoder::_str2temp(const String& temp) {
     int temp_int = temp.toInt();
     return temp_int;
 }
 
-int WeatherDecoder::_str2hum(const String& humidity) {
+int OswAppWeather::WeatherDecoder::_str2hum(const String& humidity) {
     int hum = humidity.toInt();
     hum = (hum*10)+ 5;
     return hum;
 }
 
-int WeatherDecoder::_str2pres(const String& pressure) {
+int OswAppWeather::WeatherDecoder::_str2pres(const String& pressure) {
     int pres = pressure.toInt();
     pres = pres + 850;
     return pres;
 }
 
-int WeatherDecoder::_str2wthr(const String& weather) {
+int OswAppWeather::WeatherDecoder::_str2wthr(const String& weather) {
     char wthr = weather[0];
     return wthr - 65;
 }
 
+OswAppWeather::WeatherParser::WeatherParser() {}
 
-class WeatherParser {
-  public:
-    WeatherParser();
-    String encodeWeather(DynamicJsonDocument& doc);
-  private:
-    int _getWCond(int weather_code);
-    int cnt;
-    std::vector<OswAppWeather::weather_update_t> updates;
-    std::vector<int>clearCode{800};//0
-    std::vector<int>cloudsMin{801};//1
-    std::vector<int>cloudsMed{802};//2
-    std::vector<int>cloudsHigh{803, 804};//3
-    std::vector<int>mist{701};//4
-    std::vector<int>fog{741};//5
-    std::vector<int>snowMin{611, 612, 615, 616};//6
-    std::vector<int>snowMed{600, 613, 601, 620};//7
-    std::vector<int>snowHigh{602, 621, 622};//8
-    std::vector<int>rainMin{500, 300, 301, 302, 310, 311, 312, 313, 314, 321};//9
-    std::vector<int>rainMed{502, 501};//10
-    std::vector<int>rainHigh{503, 504, 511, 520, 521, 522, 531};//11
-    std::vector<int>thunderstorm{200, 201, 210, 211, 231, 230};//12
-    std::vector<int>thunderstormHeavy{202, 212, 221, 232};//13
-    std::vector<int>squallTornado{771, 781};//14
-    //15 ->unknown
-    std::vector<std::vector<int>>weather_conditions{clearCode, cloudsMin, cloudsMed, cloudsHigh, mist, fog, snowMin, snowMed,
-            snowHigh, rainMin, rainMed, rainHigh, thunderstorm,
-            thunderstormHeavy, squallTornado };
-};
-
-WeatherParser::WeatherParser() {}
-
-String WeatherParser::encodeWeather(DynamicJsonDocument& doc) {
+std::optional<String> OswAppWeather::WeatherParser::encodeWeather(DynamicJsonDocument& doc) {
     const char* code = nullptr;
     code = doc["cod"];
     if(strcmp("200",code)) {
         if(code==nullptr) {
             OSW_LOG_E("Error, corrupted API response.");
-            return  "ERROR_CORRUPTED_RESPONSE";
+            return std::nullopt; // ERROR_CORRUPTED_RESPONSE
         } else {
             OSW_LOG_E("Error, response code: ", code);
-            return "ERROR_API_RESPONSE";
+            return std::nullopt; // ERROR_API_RESPONSE
         }
     }
     cnt = doc["cnt"];
@@ -170,7 +124,7 @@ String WeatherParser::encodeWeather(DynamicJsonDocument& doc) {
     return encoder.getEncoded();
 }
 
-int WeatherParser::_getWCond(int weather_code) {
+int OswAppWeather::WeatherParser::_getWCond(int weather_code) {
     for(int i=0; i<15; i++) {
         for(int j=0; j < weather_conditions[i].size(); j++) {
             if(weather_code == weather_conditions[i][j]) {
@@ -278,6 +232,7 @@ void OswAppWeather::weatherRequest() {
 }
 
 bool OswAppWeather::_request() {
+    // TODO cleanup the error handling (especially the return false; statements)
     HTTPClient http;
     int code = 0;
     OSW_LOG_D("Request: ");
@@ -295,36 +250,38 @@ bool OswAppWeather::_request() {
     http.end();
     OswServiceAllTasks::wifi.disconnectWiFi();
     OSW_LOG_D("Request returned code: ", code);
-    if (code == 200) {
-        DynamicJsonDocument doc(16432);
-        deserializeJson(doc,http.getStream());
-        WeatherParser pars;
-        String encoded = pars.encodeWeather(doc);
-        int encoded_len = encoded.length();
-        char encoded_arr[encoded_len + 1];
-        strcpy(encoded_arr, encoded.c_str());
-        String encoded_S = String(encoded_arr);
-        if (!this->pref.putString("wtr",encoded_S)) {
-            OSW_LOG_E("Unable to write to NVS?!");
-            this->dataLoaded = false;
-            return false;
-        }
-    } else {
+    if (code != 200) {
         OSW_LOG_E("Unexpected API response: ", code);
+        this->dataLoaded = false;
+        return false;
+    }
+    DynamicJsonDocument doc(16432);
+    deserializeJson(doc,http.getStream());
+    WeatherParser pars;
+    std::optional<String> encoded = pars.encodeWeather(doc);
+    if(!encoded.has_value()) {
+        OSW_LOG_E("Something went wrong with the encoding of the weather data?!");
+        this->dataLoaded = false;
+        return false;
+    }
+    int encoded_len = encoded.value().length();
+    char encoded_arr[encoded_len + 1];
+    strcpy(encoded_arr, encoded.value().c_str());
+    String encoded_S = String(encoded_arr);
+    if (!this->pref.putString("wtr",encoded_S)) {
+        OSW_LOG_E("Unable to write weather-update to NVS?!");
         this->dataLoaded = false;
         return false;
     }
     this->requestMode=false;
     bool res = this->loadData();
-    if (res) {
-        OSW_LOG_D("weather updated correctly");
-        this->dataLoaded=true;
-        return true;
-    } else {
+    if (!res) {
         this->dataLoaded = false;
         return false;
     }
-
+    OSW_LOG_D("weather updated correctly");
+    this->dataLoaded=true;
+    return true;
 }
 #else
 void OswAppWeather::weatherRequest() {
